Massive refactor of configuration form
Put all functions inside a custom hook, removed functionality related to setting all configuration (we are not doing this anymore). Looks better I would say and all still works
